% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCI.R
\name{GCI}
\alias{GCI}
\title{Plant stress}
\usage{
GCI(sentinel_image, green_band, nir_band, output_path = NULL)
}
\arguments{
\item{sentinel_image}{Character string representing the file name of the raster stack that contains the image data (sentinel stacked image)}

\item{green_band}{Integer, representing the band numbers to be used for the green band}

\item{nir_band}{Integer, representing the band numbers to be used for the nir band}

\item{output_path}{Optional character string, representing the file path where the calculated GCI raster will be exported}
}
\value{
GCI raster
}
\description{
Plant stress
}
