% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_image.R
\name{load_image}
\alias{load_image}
\title{Load Image}
\usage{
load_image(file_path)
}
\arguments{
\item{file_path}{The path to the image file}
}
\value{
A raster stack representing the loaded image
}
\description{
Load Image
}
