% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDMI.R
\name{NDMI}
\alias{NDMI}
\title{Plant water stress}
\usage{
NDMI(sentinel_image, nir_band, swir_band, output_path = NULL)
}
\arguments{
\item{sentinel_image}{Character string representing the file name of the raster stack that contains the image data (sentinel stacked image)}

\item{nir_band}{Integer, representing the band numbers to be used for the nir band}

\item{swir_band}{Integer, representing the band numbers to be used for the swir band}

\item{output_path}{Optional character string, representing the file path where the calculated NDMI raster will be exported}
}
\value{
NDMI raster
}
\description{
Plant water stress
}
