% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDRE.R
\name{NDRE}
\alias{NDRE}
\title{Nitrogen Content Map}
\usage{
NDRE(sentinel_image, nir_band, rededge_band, output_path = NULL)
}
\arguments{
\item{sentinel_image}{character string representing the file name of the raster stack that contains the image data (sentinel stacked image)}

\item{nir_band}{integer, representing the band numbers to be used for the nir band}

\item{rededge_band}{integer, representing the band numbers to be used for the red edge band}

\item{output_path}{optional character string, representing the file path where the calculated NDRE raster will be exported}
}
\value{
NDRE raster
}
\description{
Nitrogen Content Map
}
