% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_FCC_Image.R
\name{Plot_FCC_Image}
\alias{Plot_FCC_Image}
\title{Plotting Image in False Color Combination}
\usage{
Plot_FCC_Image(image_file, r_band, g_band, b_band)
}
\arguments{
\item{image_file}{Character string representing the file name of the raster stack that contains the image data.}

\item{r_band}{Integer, representing the band numbers to be used for the red}

\item{g_band}{Integer, representing the band numbers to be used for the green}

\item{b_band}{Integer, representing the band numbers to be used for the blue}
}
\value{
Plot the raster according to given band combinations
}
\description{
Plotting Image in False Color Combination
}
